Remove devtools
